apply certik audit report
